requirements.txt
hikari-lightbulb==3.0.0a1
